HFP-120 [feat] 프로필 파일 작업 및 리뷰 (#223)
* HFP-120 [fix] 프로필 preview 오류 해결
* HFP-120 [feat] 마이페이지 프리랜서 (고용주 리뷰) 점수 API 구현)
* HFP-120 [feat] 마이페이지 S3 파일 업로드 구현
- 프로필 사진 업로드
- 프리랜서 포트폴리오 업로드
* HFP-120 [feat] 포트폴리오 template 설정

+++ freebridge/user/src/main/java/com/fallguys/mypage/entity/freelancer/Expertise.java
@@ -11,7 +11,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class Expertise {
-    private Integer programming;
-    private Integer framework;
-    private Integer problemSolving;
+    private Double programming;
+    private Double framework;
+    private Double problemSolving;
 }
